make dto names uniform and move into common folder
