+++ src/main/java/co/uk/bransby/equinetrainingtrackerapi/controllers/EquineController.java
@@ -1,6 +1,6 @@
 package co.uk.bransby.equinetrainingtrackerapi.controllers;
 
-import co.uk.bransby.equinetrainingtrackerapi.dtos.EquineDto;
+import co.uk.bransby.equinetrainingtrackerapi.models.dto.EquineDto;
 import co.uk.bransby.equinetrainingtrackerapi.models.Equine;
 import co.uk.bransby.equinetrainingtrackerapi.services.EquineService;
 import org.modelmapper.ModelMapper;
