+++ src/main/java/co/uk/bransby/equinetrainingtrackerapi/models/dto/CategoryDto.java
@@ -4,10 +4,10 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 @Data
-public class DisruptionDTO {
+public class CategoryDto {
     Long id;
     String name;
 }
